refactor(harness): orchestrator service over adapters; retire claude-only modules
